+++ jobs/tests/test_inventory_to_records_functional.py
@@ -18,17 +18,17 @@ class CsvToRecordsTest(asynctest.TestCase):
         self.argv = sys.argv
 
         filename = os.path.join(here, 'data', 'inventory-simple.csv')
-        sys.stdin = open(filename, "r")
+        sys.stdin = open(filename, 'r')
         sys.stdout = io.StringIO()
-        sys.argv = ["inventory-to-records"]
+        sys.argv = ['inventory-to-records']
 
         async def fake_stream(loop, stream):
             # Workaround for "Pipe transport is for pipes/sockets only".
             # Here stream is just a file descriptor.
             for line in stream.readlines():
-                yield bytes(line, "utf-8")
+                yield bytes(line, 'utf-8')
 
-        p = mock.patch("buildhub.inventory_to_records.stream_as_generator",
+        p = mock.patch('buildhub.inventory_to_records.stream_as_generator',
                        wraps=fake_stream)
         self.addCleanup(p.stop)
         p.start()
@@ -45,7 +45,7 @@ class CsvToRecordsTest(asynctest.TestCase):
         output = sys.stdout.getvalue()
         records = [json.loads(o) for o in output.split('\n') if o]
         assert records == [{
-            "data": {
+            'data': {
                 'id': 'firefox_nightly_2017-05-15-10-02-38_55-0a1_linux-x86_64_en-us',
                 'build': {
                     'id': '20170515100238',
@@ -83,7 +83,7 @@ class CsvToRecordsTest(asynctest.TestCase):
                 }
             }
         }, {
-            "data": {
+            'data': {
                 'id': 'firefox_52-0_linux-x86_64_fr',
                 'build': {
                     'id': '20170302120751',
@@ -178,32 +178,32 @@ class CsvToRecordsTest(asynctest.TestCase):
                 }
             }
         }, {
-            "data": {
-                "id": "devedition_aurora_55-0b1rc5_win64_pt-br",
-                "source": {
-                    "product": "devedition",
-                    "revision": "6872377277a618b2b9e0d2b4c2b9e51765ac199e",
-                    "repository": "https://hg.mozilla.org/releases/mozilla-beta",
-                    "tree": "releases/mozilla-beta"
+            'data': {
+                'id': 'devedition_aurora_55-0b1rc5_win64_pt-br',
+                'source': {
+                    'product': 'devedition',
+                    'revision': '6872377277a618b2b9e0d2b4c2b9e51765ac199e',
+                    'repository': 'https://hg.mozilla.org/releases/mozilla-beta',
+                    'tree': 'releases/mozilla-beta'
                 },
-                "target": {
-                    "platform": "win64",
-                    "os": "win",
-                    "locale": "pt-BR",
-                    "version": "55.0b1rc5",
-                    "channel": "aurora"
+                'target': {
+                    'platform': 'win64',
+                    'os': 'win',
+                    'locale': 'pt-BR',
+                    'version': '55.0b1rc5',
+                    'channel': 'aurora'
                 },
-                "download": {
-                    "url": "https://archive.mozilla.org/pub/devedition/candidates/"
-                           "55.0b1-candidates/build5/win64/pt-BR/firefox-55.0b1.zip",
-                    "mimetype": "application/zip",
-                    "size": 53718907,
-                    "date": "2017-06-14T00:41:56Z"
+                'download': {
+                    'url': 'https://archive.mozilla.org/pub/devedition/candidates/'
+                           '55.0b1-candidates/build5/win64/pt-BR/firefox-55.0b1.zip',
+                    'mimetype': 'application/zip',
+                    'size': 53718907,
+                    'date': '2017-06-14T00:41:56Z'
                 },
-                "build": {
-                    "id": "20170612224034",
-                    "date": "2017-06-12T22:40:34Z",
-                    "number": 5,
+                'build': {
+                    'id': '20170612224034',
+                    'date': '2017-06-12T22:40:34Z',
+                    'number': 5,
                     'as': 'ml64.exe',
                     'cc': 'c:/builds/moz2_slave/m-beta-w64-de-0000000000000000/build/'
                           'src/vs2015u3/VC/bin/amd64/cl.exe',
@@ -212,6 +212,45 @@ class CsvToRecordsTest(asynctest.TestCase):
                     'date': '2017-06-12T22:40:34Z',
                     'host': 'x86_64-pc-mingw32',
                     'target': 'x86_64-pc-mingw32'
+                }
+            }
+        }, {
+            'data': {
+                'id': 'firefox_beta_51-0b11_macosx-eme-free_mr',
+                'build': {
+                    'as': '$(CC)',
+                    'cc': '/usr/local/bin/ccache '
+                          '/builds/slave/m-beta-m64-0000000000000000000/build/src/clang/bin/clang '
+                          '-arch x86_64 -std=gnu99',
+                    'cxx': '/usr/local/bin/ccache '
+                           '/builds/slave/m-beta-m64-0000000000000000000/build/src/clang/bin/'
+                           'clang++ -arch x86_64 -std=gnu++11',
+                    'date': '2017-01-03T03:11:19Z',
+                    'host': 'x86_64-apple-darwin11.2.0',
+                    'id': '20170103031119',
+                    'ld': 'ld',
+                    'number': 1,
+                    'target': 'x86_64-apple-darwin11.2.0'
+                },
+                'download': {
+                    'date': '2017-01-04T00:05:18Z',
+                    'mimetype': 'application/x-apple-diskimage',
+                    'size': 85984611,
+                    'url': 'https://archive.mozilla.org/pub/firefox/releases/51.0b11/mac-EME-free/'
+                           'mr/Firefox 51.0b11.dmg'
+                },
+                'source': {
+                    'product': 'firefox',
+                    'repository': 'https://hg.mozilla.org/releases/mozilla-beta',
+                    'revision': '0a17d39220700e742bf37a960967480b2f8159f1',
+                    'tree': 'releases/mozilla-beta'
+                },
+                'target': {
+                    'channel': 'beta',
+                    'locale': 'mr',
+                    'os': 'mac',
+                    'platform': 'macosx-eme-free',
+                    'version': '51.0b11'
                 }
             }
         }]