+++ jobs/CHANGELOG.rst
@@ -6,7 +6,7 @@ This document describes changes between each past release.
 0.5.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix Mac OS X metadata URLs (fixes #261)
 
 
 0.4.1 (2017-09-29)