+++ jobs/CHANGELOG.rst
@@ -8,6 +8,7 @@ This document describes changes between each past release.
 
 - Fix Mac OS X metadata URLs (fixes #261)
 - Skip exe installers where version is missing from URL (fixes #263)
+- Fix Fennec metadata location (fixes #264)
 
 
 0.4.1 (2017-09-29)