+++ jobs/CHANGELOG.rst
@@ -12,6 +12,8 @@ This document describes changes between each past release.
 - Fix beta and devedition medata URLs (#269)
 - Skip exe installers where version is missing from URL (fixes #263)
 - Fix Fennec metadata location (fixes #264)
+- Fix caching when partial updates metadata is missing (fixes #276)
+- Fix handling of bad server response, like XML (fixes #259)
 
 
 0.4.1 (2017-09-29)